feat: add trip creation modal component
